+++ .storybook/theme.ts
@@ -4,6 +4,6 @@ export default create({
   base: 'light',
   brandTitle: 'Gov UK Design system (React)',
   brandUrl: 'https://github.com/rottitime/govuk-design-react',
-  // brandImage: 'https://storybook.js.org/images/placeholders/350x150.png',
+  brandImage: 'https://rottitime.github.io/govuk-design-react/doc/logo2.webp',
   brandTarget: '_self'
 })